历史库里的MNode、HNode、UpdateHNode(存储过程)里的real类型，都改成float双精度

+++ Rcm.Site/Extensions/Common.cs
@@ -121,7 +121,7 @@ namespace Rcm.Site.Extensions {
             }
         }
 
-        public static string GetValueDisplay(Rcm.Core.Domain.Point current, float value) {
+        public static string GetValueDisplay(Rcm.Core.Domain.Point current, double value) {
             switch(current.Type) {
                 case EnmScType.Dic:
                 case EnmScType.Doc:
